+++ bindings/python/python/decibri/__init__.py
@@ -33,6 +33,7 @@ from decibri.exceptions import (
     DeviceFailed,
     DeviceIndexOutOfRange,
     FileConsumed,
+    FileEngaged,
     FileReadFailed,
     ForkAfterOrtInit,
     FramesPerBufferOutOfRange,